# docs/swagger.yaml
definitions:
  main.Device:
    properties:
      appium_port:
        type: integer
      device_name:
        type: string
      device_os_version:
        type: string
      device_udid:
        type: string
      wda_mjpeg_port:
        type: integer
      wda_port:
        type: integer
    type: object
  main.ErrorJSON:
    properties:
      error_message:
        type: string
      event:
        type: string
    type: object
  main.IOSAppInstall:
    properties:
      ipa_name:
        type: string
    type: object
  main.IOSAppUninstall:
    properties:
      bundle_id:
        type: string
    type: object
  main.IOSDeviceInfo:
    properties:
      deviceConfig:
        $ref: '#/definitions/main.Device'
      installedAppsBundleIDs:
        items:
          type: string
        type: array
    type: object
  main.ProjectConfig:
    properties:
      devices_host:
        type: string
      selenium_hub_host:
        type: string
      selenium_hub_port:
        type: string
      selenium_hub_protocol_type:
        type: string
      wda_bundle_id:
        type: string
    type: object
  main.SimpleResponseJSON:
    properties:
      event:
        type: string
      message:
        type: string
    type: object
  main.SudoPassword:
    properties:
      sudo_password:
        type: string
    type: object
info:
  contact: {}
paths:
  /configuration/build-image:
    post:
      description: Starts building the 'ios-appium' image in a goroutine and just
        returns Accepted
      responses:
        "202":
          description: ""
      summary: Build 'ios-appium' image
      tags:
      - configuration
  /configuration/remove-image:
    post:
      description: Removes the 'ios-appium' Docker image
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.SimpleResponseJSON'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.ErrorJSON'
      summary: Remove 'ios-appium' image
      tags:
      - configuration
  /configuration/remove-ios-listener:
    post:
      description: Deletes udev rules from /etc/udev/rules.d and reloads udev
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.SimpleResponseJSON'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.ErrorJSON'
      summary: Removes iOS device listener
      tags:
      - configuration
  /configuration/set-sudo-password:
    put:
      consumes:
      - application/json
      description: Sets your sudo password in ./env.json. The password is needed for
        operations requiring elevated permissions like setting up udev.
      parameters:
      - description: Sudo password value
        in: body
        name: config
        required: true
        schema:
          $ref: '#/definitions/main.SudoPassword'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.SimpleResponseJSON'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.ErrorJSON'
      summary: Set sudo password
      tags:
      - configuration
  /configuration/setup-ios-listener:
    post:
      description: Creates udev rules, moves them to /etc/udev/rules.d and reloads
        udev. Copies usbmuxd.service to /lib/systemd/system and enables it
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.SimpleResponseJSON'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.ErrorJSON'
      summary: Sets up iOS device listener
      tags:
      - configuration
  /configuration/update-config:
    put:
      consumes:
      - application/json
      description: Updates one  or multiple configuration values
      parameters:
      - description: Update config
        in: body
        name: config
        required: true
        schema:
          $ref: '#/definitions/main.ProjectConfig'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.SimpleResponseJSON'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.ErrorJSON'
      summary: Update project configuration
      tags:
      - configuration
  /containers/{container_id}/logs:
    post:
      description: Get logs of container by providing container ID
      parameters:
      - description: Container ID
        in: path
        name: container_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.SimpleResponseJSON'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.ErrorJSON'
      summary: Get container logs
      tags:
      - containers
  /containers/{container_id}/remove:
    post:
      description: Removes container by provided container ID
      parameters:
      - description: Container ID
        in: path
        name: container_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.SimpleResponseJSON'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.ErrorJSON'
      summary: Remove container
      tags:
      - containers
  /containers/{container_id}/restart:
    post:
      description: Restarts container by provided container ID
      parameters:
      - description: Container ID
        in: path
        name: container_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.SimpleResponseJSON'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.ErrorJSON'
      summary: Restart container
      tags:
      - containers
  /device-logs/{log_type}/{device_udid}:
    get:
      description: Get logs by type
      parameters:
      - description: Log Type
        in: path
        name: log_type
        required: true
        type: string
      - description: Device UDID
        in: path
        name: device_udid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.SimpleResponseJSON'
      summary: Get logs for iOS device container
      tags:
      - device-logs
  /ios-devices/{device_udid}/info:
    get:
      description: Get info for an iOS device - installed apps, Appium config
      parameters:
      - description: Device UDID
        in: path
        name: device_udid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.IOSDeviceInfo'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.ErrorJSON'
      summary: Get info for iOS device
      tags:
      - ios-devices
  /ios-devices/{device_udid}/install-app:
    post:
      description: Installs *.ipa or *.app from the './ipa' folder
      parameters:
      - description: Device UDID
        in: path
        name: device_udid
        required: true
        type: string
      - description: Install iOS app
        in: body
        name: config
        required: true
        schema:
          $ref: '#/definitions/main.IOSAppInstall'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.SimpleResponseJSON'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.ErrorJSON'
      summary: Install app on iOS device
      tags:
      - ios-devices
  /ios-devices/{device_udid}/uninstall-app:
    post:
      description: Uninstalls app from iOS device by provided bundleID
      parameters:
      - description: Device UDID
        in: path
        name: device_udid
        required: true
        type: string
      - description: Uninstall iOS app
        in: body
        name: config
        required: true
        schema:
          $ref: '#/definitions/main.IOSAppUninstall'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.SimpleResponseJSON'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.ErrorJSON'
      summary: Uninstall app from iOS device
      tags:
      - ios-devices
  /ios_containers/update:
    post:
      description: Creates (or removes respectively) iOS containers based on the connected
        and registered devices
      parameters:
      - description: Device UDID
        in: path
        name: device_udid
        required: true
        type: string
      responses:
        "202":
          description: ""
      summary: Update iOS containers
      tags:
      - ios-devices
  /project-logs:
    get:
      description: Provides project logs as plain text response
      responses:
        "200":
          description: ""
      summary: Get project logs
      tags:
      - project-logs
swagger: "2.0"